redesign for SecretProvider interface
