+++ lib/addons/secrets-store/secret-provider.ts
@@ -1,0 +1,20 @@
+import { KubernetesManifest } from "@aws-cdk/aws-eks";
+import { ClusterInfo } from "../../stacks/cluster-types";
+import { ApplicationTeam } from "../../teams";
+
+/**
+ * Secret Provider Interface
+ * You can provide() your own Secrets driver OR
+ * use the one we provide using AWS Secrets Manager and Config Provider(ASCP)
+ */
+export interface SecretsProvider {
+  provide(clusterInfo: ClusterInfo): KubernetesManifest;
+}
+
+/**
+ * Secrets to be provided
+ */
+export interface Secrets {
+  secrets: any;
+  setupSecrets?(clusterInfo: ClusterInfo, team: ApplicationTeam): any;
+}
